+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Implemented text justify for *rich text* composed of `Wrap!` and `Text!`.
 
 # 0.13.11